[Products API slice and GetProducts endpoint] create our productsApiSlice.js in our slices folder.
The code starts by importing two constants: PRODUCTS_URL and apiSlice. The PRODUCTS_URL constant contains the URL of the API endpoint that returns a list of products, while the apiSlice is a reference to the Redux Toolkit API slice that is used to manage the state of the API requests.

The productsApiSlice is then created by calling the injectEndpoints method on the apiSlice and passing a configuration object that defines the endpoints for the slice. In this case, the slice has only one endpoint, getProducts, which is defined using the builder.query method.

The query method takes a configuration object that specifies the query function to be used to fetch data from the API endpoint. In this case, the query function is a simple function that returns the PRODUCTS_URL constant.

The keepUnusedDataFor property is used to specify the number of requests to keep in the cache, after which the oldest requests will be removed. This property is optional and can be omitted if you don't need to specify a cache size.

Finally, the useGetProductsQuery hook is created by exporting the productsApiSlice and its associated hook. This allows components in the application to access the data returned from the API endpoint using the useGetProductsQuery hook.

Next we refactored screens/Homescreen.jsx by first removing the axios and useEffect&&useState hook imports and deleting the associated code above the return statement. Then we import our useGetProductsQuery hook from productsApiSlice that we created and use this in place of our deleted code prior to refactoring the code in the return statment to reflect our current states.

Additionally fixed a bug in constants.js that was disabling our ability to use the proxy we set up earlier.