[Loader and Message Components] Added Loader jsx to components folder. here we  created a spinner component to provide a visual indicator that our app is loading.
The Spinner component displays a loading animation with a border and a spinner.

The Spinner component takes several props, including animation, size, and variant. In this case, the animation prop is set to "border", which displays a border around the spinner. The size prop is set to "100px" for the width and height, and the margin is set to "auto" to center the spinner. The display property is set to "block" to make the spinner visible.

The Spinner component is wrapped in a div element with a fixed width and height of "100px". The margin is set to "auto" to center the spinner within the div.

Our loader component is then imported in homeScreen.jsx and productScreen.jsx.

Next we created Message.jsx in our components folder. a basic alert component that takes in a prop called variant, which determines the appearance of the alert. The default value for this prop is "info". The component returns an Alert component from the react-bootstrap library, which displays the message passed to it as its children. This component was then used to wrap our error mesages in HomeScreen.jsx and ProductScreen.jsx and given the varient danger to indicate that this is an error message.