[Shopping Cart Functionality] [Cart Slice and Reducer] create ne file (cartSlice.js) in src/slices. import createSlice from reactjs/toolkit. Using the CreateSlice function from the @reduxjs/toolkit library to create a new slice of the application state.
The initialState variable defines the initial state of the cart slice, which is retrieved from local storage if it exists, or set to an empty array if it does not.

The cartSlice variable is then created using the CreateSlice function, with the name "cart", the initialState variable, and an empty reducers object.

Finally, the default export is set to the cartSlice.reducer, which is the reducer function for the cart slice.

Next in our store.js import our cartSliceReducer from our slices directory and add our cart to our reducer: {} in our configureStore.  This slice is a way to split the state of the store into smaller, more manageable pieces. In this case, the "cart" slice is responsible for managing the state of the user's cart.