[Add to Cart Function ] Add addToCartFunction inside of our cartslice reducer.
The addToCart reducer handles adding a new item to the cart. It takes in the current state and an action object that contains the payload of the new item to be added to the cart.

The code first checks if the new item already exists in the cart. If it does, it updates the existing item in the state by replacing it with the new item. If the new item does not exist, it adds it to the cart.

Next, the code calculates the total price of the items in the cart by using the reduce method to sum up the price of each item multiplied by its quantity. It then updates the itemsPrice property of the state with the calculated value.

The code then calculates the shipping price based on the total price of the items in the cart. If the total price is greater than $100, the shipping price is set to $0. Otherwise, it is set to $10.

The code then calculates the tax price based on the total price of the items in the cart. The tax rate is set to 15%.

Finally, the code calculates the total price by adding the items price, shipping price, and tax price. It then rounds the total price to two decimal places and updates the totalPrice property of the state with the calculated value.

The code then saves the updated cart state to local storage using the JSON.stringify method.