[Quantity and Add To Cart Function] in Product Screen.jsx import useState from react. Then  initialize a React state variable called qty with a default value of 1. This state variable is used to keep track of the quantity of items the user wants to add to their cart.
next add  conditional rendering of the Qty select input field. It is only rendered if the countInStock property of the product object is greater than 0, indicating that the product is in stock.

The code uses a React hook, useState, to manage the state of the Qty select input field. The current value of the Qty select input field is stored in the qty state variable. The onChange event handler updates the qty state variable with the value selected in the Qty select input field.

The code uses the map method to create an array of numbers from 0 to the product countInStock value, and then uses the key and value attributes to create the options for the Qty select input field. The selected option is highlighted in the Qty select input field based on the value of the qty state variable.

Add onClick event to the button component to make a call to our addToCartHandler which we are to create next.

The selected code is a function that is used to add an item to the cart. It is defined inside the ProductScreen component and is called when the Add to Cart button is clicked.

The addToCartHandler function first dispatches an action creator called addToCart, which takes an object with the product details and the quantity as its payload. This action will be handled by the cart reducer, which will update the cart state in the store.

After dispatching the addToCart action, the navigate function is used to redirect the user to the cart page.

The addToCartHandler is used to add an item to the cart, regardless of whether the item is in stock or not. However the add to cart button is disabled if there are 0 items in stock thus no errors will appear.