[Add Item Count To Header] in frontend/src/components/header.js import useSelector from react-redux. we use this to select the cartItems property from the state object that is returned by the useSelector hook. This means that the component will re-render whenever the cart state changes, and it will get the updated cart items.
In this case, the Header component is using the cartItems property to determine whether there are any items in the cart and to display the cart icon with the number of items in the cart.

Finally we add this next to our shopping cart component in the app header using conditional rendering of a React Bootstrap Badge component. The condition is checking if the length of the cartItems array is greater than 0. If it is, a Badge component using its pill property to give it a round appearance, with a success variant and the sum of all the quantities of the items in the cart is rendered.

The Badge component is imported from React Bootstrap and is used to display a small, informational badge. The success variant is used to indicate a positive or successful status. The className prop is used to add custom CSS classes to the component.

The sum of all the quantities of the items in the cart is calculated using the reduce method. The reduce method takes a callback function that is executed for each element in the array, accumulating the return value of the callback into a single value. In this case, the callback function adds the current quantity of each item to the accumulator variable a. The initial value of the accumulator is 0.

Overall, the code is checking if there are any items in the cart, and if there are, it is rendering a Badge component with the total number of items in the cart.